% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrammer.R, R/includeGraphic.R
\name{includeGraphic.grViz}
\alias{includeGraphic.grViz}
\alias{includeGraphic.dgr_graph}
\alias{includeGraphic}
\alias{includeGraphic.default}
\title{Includes a graphic into a RMarkdown document}
\usage{
\method{includeGraphic}{grViz}(gr, nr = 0)

\method{includeGraphic}{dgr_graph}(gr, nr = 0)

includeGraphic(gr, nr = 0)

\method{includeGraphic}{default}(gr, nr = 0)
}
\arguments{
\item{gr}{graph}

\item{nr}{optional number when creating more than one graph in a chunk}
}
\description{
Graphs created in a loop or by some packages like Diagrammer
are not always automatically included in the final document
produced from a RMarkdown document. This function helps including
graphics in documents.
}
