% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showgraphs.R
\name{showGraph}
\alias{showGraph}
\title{Includes graphs produced in a loop into final document}
\usage{
showGraph(gr, withBlankLine = TRUE)
}
\arguments{
\item{gr}{graph}

\item{withBlankLine}{if TRUE, then a blank line is printed}
}
\description{
When producing graphics within a loop, only the last
graph will be included by default. When calling this
function, the graph is printed. If withBlanLine is TRUE,
then a blank line is printed, causing the creation of
figure environments.
}
