% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{texEnv}
\alias{texEnv}
\title{Starts a special LaTeX environment}
\usage{
texEnv()
}
\description{
Some environments like \code{align} must not be
included within \verb{$$}..\verb{$$} when creating a PDF.
For Word documents, however, the \verb{$$} are needed.
}
