% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdutil.R
\docType{package}
\name{rmdutil}
\alias{rmdutil}
\title{Utility functions to simplify creation of RMarkdown-Documents}
\description{
Functions to:
\itemize{
\item Include graphs created by a loop into documents
\item Include DiagrammeR graphs in PDF
}
}
